Update README.md
In the given task, two classes, Student and StudentManagementSystem are created respectively.In Student class fields of ID,name,marks are initialized. In StudentManagementSystem class the fields and methods present in the Student class are called and the CLI based objects such as - addStudent,viewStudent, updateStudent,deleteStudent are created. The objects are then stored in a data structure ArrayList.